refactor(src): replaces got with axios, package is now compatible with both the browser and nodejs
This update abstracts the http request method out a little futher than it was in a "fetcher" object
and includes a utility wrapper for conforming request responses to the ImgurApiResponse model. All
test have been updated and support for automatic filepath uploads has been dropped in favor of
cross-platform support.

BREAKING CHANGE: The package can no longer be used with string path variables pointing to local
files. This moves a single line of code, using fs.createReadStream, out of the project and places
that responsibility on the consumer. The package now supports streams in favor of the string
filepath upload.